requirements.txt
# Core Dependencies
Pillow>=10.4.0

# PDF-Generierung
reportlab>=4.2.5

# SVG-Verarbeitung (Wand = ImageMagick Python-Wrapper)
Wand>=0.6.13

# GUI Framework
PyQt6>=6.7.0
PyQt6-Qt6>=6.7.0

# Optional (für spätere Features)
openpyxl>=3.1.5  # Excel-Import (für Batch-Verarbeitung)
# PyInstaller>=6.0.0  # EXE-Build
